DTSCCI-680 Refactor CreateClaimSpecCallbackHandler
